Renamed landmark_parser to landmark_manager
