+++ localization/landmark_based_localizer/landmark_manager/CMakeLists.txt
@@ -1,5 +1,5 @@
 cmake_minimum_required(VERSION 3.14)
-project(landmark_parser)
+project(landmark_manager)
 
 find_package(autoware_cmake REQUIRED)
 autoware_package()
@@ -12,8 +12,8 @@ endif()
 find_package(ament_cmake_auto REQUIRED)
 ament_auto_find_build_dependencies()
 
-ament_auto_add_library(landmark_parser
-  src/landmark_parser_core.cpp
+ament_auto_add_library(landmark_manager
+  src/landmark_manager.cpp
 )
 
 ament_auto_package(
